pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot template using Poetry, Docker and GitHub Actions"
authors = ["Joey Harrison <joeyh021@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.1.0"
pydantic = "^1.10.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.1.0"
mypy = "^0.991"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]
